fix unit test for all user service methods

+++ src/services/user.service.ts
@@ -16,10 +16,10 @@ export const AddUser = async (user: User): Promise<TypeORMResponse.Signature> =>
    }
 };
 export const getUserById = async (userId: string): Promise<TypeORMResponse.Signature> => {
-  try {
+   try {
       if (!userId) return formatResponse<TypeORMResponse.RecordNotFound>({ message: 'NotAcceptable: No UserId provided', statusCode: 406 });
       const user = await AppDataSource.createQueryBuilder().select('U').from(User, 'U').where('U.userId = :id', { id: userId }).getOne();
-    if (!user) {
+      if (!user) {
          return formatResponse<TypeORMResponse.RecordNotFound>({
             queryIdentifier: userId,
             message: `User with ID ${userId} not found`,
@@ -33,7 +33,8 @@ export const getUserById = async (userId: string): Promise<TypeORMResponse.Signa
 };
 /* Late we will think about if we need all users*/
 export const deleteUser = async (userId: string): Promise<TypeORMResponse.Signature> => {
-   try {
+  try {
+      if (!userId) return formatResponse<TypeORMResponse.RecordNotFound>({ message: 'NotAcceptable: No UserId provided', statusCode: 406 });
       const deletedResult = await AppDataSource.createQueryBuilder().delete().from(User).where('userid= :userid', { userid: userId }).execute();
       if (!deletedResult.affected) {
          return formatResponse<TypeORMResponse.RecordNotFound>({
@@ -47,14 +48,15 @@ export const deleteUser = async (userId: string): Promise<TypeORMResponse.Signat
       throw new Error(error);
    }
 };
-export const updateUser = async (updateUserData: User, userid: string): Promise<TypeORMResponse.Signature> => {
-   try {
-      const updatedResults = await AppDataSource.createQueryBuilder().update(User).set(updateUserData).where('userid= :userid', { userid: userid }).execute();
+export const updateUser = async (updateUserData: User, userId: string): Promise<TypeORMResponse.Signature> => {
+  try {
+      if (!userId) return formatResponse<TypeORMResponse.RecordNotFound>({ message: 'NotAcceptable: No UserId provided', statusCode: 406 });
+      const updatedResults = await AppDataSource.createQueryBuilder().update(User).set(updateUserData).where('userid= :userid', { userid: userId }).execute();
       if (!updatedResults.affected) {
          return formatResponse<TypeORMResponse.RecordNotFound>({
-            queryIdentifier: userid,
+            queryIdentifier: userId,
             statusCode: 404,
-            message: `No record was updated for user: ${userid}`,
+            message: `No record was updated for user: ${userId}`,
          });
       }
       return formatResponse<UpdateResult>(updatedResults);
